Update text component
