+++ packages/terra-text/src/Text.module.scss
@@ -50,7 +50,7 @@
 
   .font-stack {
     &::after {
-      content: '\"Helvetica Neue\", Helvetica, Arial, sans-serif';
+      content: var(--terra-text-v2-font-stack-after-content, '\"Helvetica Neue\", Helvetica, Arial, sans-serif');
       display: inline-block;
     }
   }
