--- packages/terra-text/src/terra-dev-site/test/text/TextFontSize.test.jsx
@@ -1,41 +1,0 @@
-import React from 'react';
-
-import Text from '../../../Text';
-
-const TextFontSize = () => (
-  <div>
-    <Text id="textFontSize100" fontSize={100}>Font Size 100</Text>
-    <br />
-    <Text id="textFontSize92" fontSize={92}>Font Size 92</Text>
-    <br />
-    <Text id="textFontSize72" fontSize={72}>Font Size 72</Text>
-    <br />
-    <Text id="textFontSize64" fontSize={64}>Font Size 64</Text>
-    <br />
-    <Text id="textFontSize56" fontSize={56}>Font Size 56</Text>
-    <br />
-    <Text id="textFontSize50" fontSize={50}>Font Size 50</Text>
-    <br />
-    <Text id="textFontSize46" fontSize={46}>Font Size 46</Text>
-    <br />
-    <Text id="textFontSize36" fontSize={36}>Font Size 36</Text>
-    <br />
-    <Text id="textFontSize32" fontSize={32}>Font Size 32</Text>
-    <br />
-    <Text id="textFontSize24" fontSize={24}>Font Size 24</Text>
-    <br />
-    <Text id="textFontSize20" fontSize={20}>Font Size 20</Text>
-    <br />
-    <Text id="textFontSize18" fontSize={18}>Font Size 18</Text>
-    <br />
-    <Text id="textFontSize16" fontSize={16}>Font Size 16</Text>
-    <br />
-    <Text id="textFontSize14" fontSize={14}>Font Size 14</Text>
-    <br />
-    <Text id="textFontSize12" fontSize={12}>Font Size 12</Text>
-    <br />
-    <Text id="textFontSize10" fontSize={10}>Font Size 10</Text>
-  </div>
-);
-
-export default TextFontSize;
