+++ packages/terra-text/src/terra-dev-site/test/text/TextFontSize1.test.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+import Text from '../../../Text';
+
+const TextFontSize = () => (
+  <div id="fontSize1">
+    <Text id="textFontSize10" fontSize={10}>Font Size 10</Text>
+    <br />
+    <Text id="textFontSize12" fontSize={12}>Font Size 12</Text>
+    <br />
+    <Text id="textFontSize14" fontSize={14}>Font Size 14</Text>
+    <br />
+    <Text id="textFontSize16" fontSize={16}>Font Size 16</Text>
+    <br />
+    <Text id="textFontSize18" fontSize={18}>Font Size 18</Text>
+    <br />
+    <Text id="textFontSize20" fontSize={20}>Font Size 20</Text>
+    <br />
+    <Text id="textFontSize24" fontSize={24}>Font Size 24</Text>
+    <br />
+    <Text id="textFontSize32" fontSize={32}>Font Size 32</Text>
+  </div>
+);
+
+export default TextFontSize;
