+++ packages/terra-text/src/terra-dev-site/test/text/TextFontSize3.test.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import Text from '../../../Text';
+
+const TextFontSize = () => (
+  <div id="fontSize3">
+    <Text id="textFontSize72" fontSize={72}>Font Size 72</Text>
+    <br />
+    <Text id="textFontSize92" fontSize={92}>Font Size 92</Text>
+    <br />
+    <Text id="textFontSize100" fontSize={100}>Font Size 100</Text>
+  </div>
+);
+
+export default TextFontSize;
