+++ packages/terra-text/tests/wdio/text-spec.js
@@ -1,4 +1,4 @@
-const viewports = Terra.viewports('medium');
+const viewports = Terra.viewports('huge');
 
 describe('Text', () => {
   describe('Default', () => {
@@ -22,10 +22,22 @@ describe('Text', () => {
     Terra.should.matchScreenshot({ viewports });
   });
 
-  describe('Font Size', () => {
-    before(() => browser.url('/#/raw/tests/terra-text/text/text-font-size'));
+  describe('Font Size Group 1', () => {
+    before(() => browser.url('/#/raw/tests/terra-text/text/text-font-size-1'));
 
     Terra.should.beAccessible({ viewports });
+    Terra.should.matchScreenshot({ viewports });
+  });
+
+  describe('Font Size Group 2', () => {
+    before(() => browser.url('/#/raw/tests/terra-text/text/text-font-size-2'));
+
+    Terra.should.matchScreenshot({ viewports });
+  });
+
+  describe('Font Size Group 3', () => {
+    before(() => browser.url('/#/raw/tests/terra-text/text/text-font-size-3'));
+
     Terra.should.matchScreenshot({ viewports });
   });
